python/requirements.txt
python3-indy==1.8.0
aiohttp==3.4.4
serpy==0.3.1
aiohttp-index==0.1
aiohttp-jinja2==1.0.0
msgpack
